+++ CHANGES
@@ -2490,7 +2490,7 @@ Version 1.4.2 - April 1, 2006 (No foolin'!)
   This speedup may break existing programs that use parse actions that
   have side-effects.  For this reason, packrat parsing is disabled when
   you first import pyparsing.  To activate the packrat feature, your
-  program must call the class method ParserElement.enablePackrat().  If
+  program must call the class method enablePackrat()().  If
   your program uses psyco to "compile as you go", you must call
   enablePackrat before calling psyco.full().  If you do not do this,
   Python will crash.  For best results, call enablePackrat() immediately
