+++ examples/lucene_grammar.py
@@ -9,8 +9,9 @@
 
 import pyparsing as pp
 from pyparsing import pyparsing_common as ppc
+from pyparsing.cache import enablePackrat
 
-pp.ParserElement.enablePackrat()
+enablePackrat()
 
 COLON, LBRACK, RBRACK, LBRACE, RBRACE, TILDE, CARAT = map(pp.Literal, ":[]{}~^")
 LPAR, RPAR = map(pp.Suppress, "()")
